requirements.txt
gradio
gspread
pandas
oauth2client
google-auth
